Second iteration at refactoring SurfaceOM
