+++ Models/Core/Model.cs
@@ -57,7 +57,8 @@ namespace Models.Core
         [XmlElement(typeof(Models.PMF.Plant))]
         [XmlElement(typeof(Models.PMF.OilPalm.OilPalm))]
         [XmlElement(typeof(Models.Soils.Soil))]
-        [XmlElement(typeof(Models.SurfaceOM.SurfaceOrganicMatter))]
+        [XmlElement(typeof(Models.Surface.SurfaceOrganicMatter))]
+        [XmlElement(typeof(Models.Surface.ResidueTypes))]
         [XmlElement(typeof(Models.SoluteManager))]
         [XmlElement(typeof(Models.AgPasture.Sward))]
         [XmlElement(typeof(Models.AgPasture.PastureSpecies))]
