+++ Models/Models.csproj
@@ -476,7 +476,9 @@
     <Compile Include="Soils\Water.cs" />
     <Compile Include="Soils\SoilWater.cs" />
     <Compile Include="Summary.cs" />
-    <Compile Include="SurfaceOrganicMatter.cs" />
+    <Compile Include="Surface\SurfaceOrganicMatter.cs" />
+    <Compile Include="Surface\ResidueType.cs" />
+    <Compile Include="Surface\ResidueTypes.cs" />
     <Compile Include="Tests.cs" />
     <Compile Include="Core\ISummary.cs" />
     <Compile Include="Core\Model.cs" />
