+++ Models/Soils/SoilWater.cs
@@ -7,7 +7,7 @@ using Models;
 using System.Xml.Serialization;
 using Models.PMF;
 using System.Runtime.Serialization;
-using Models.SurfaceOM;
+using Models.Surface;
 using Models.Soils.SoilWaterBackend;
 using Models.Interfaces;
 
@@ -1042,6 +1042,30 @@ namespace Models.Soils
         #region Tillage
 
         /// <summary>
+        /// 
+        /// </summary>
+        [Serializable]
+        public class TillageTypesList : Model
+        {
+            /// <summary>Gets or sets the type of the tillage.</summary>
+            /// <value>The type of the tillage.</value>
+            public List<TillageType> TillageType { get; set; }
+
+            /// <summary>Gets the tillage data.</summary>
+            /// <param name="Name">The name.</param>
+            /// <returns></returns>
+            public TillageType GetTillageData(string Name)
+            {
+                foreach (TillageType tillageType in TillageType)
+                {
+                    if (tillageType.Name == Name)
+                        return tillageType;
+                }
+                return null;
+            }
+        }
+
+        /// <summary>
         /// Tillages the specified default tillage name.
         /// </summary>
         /// <param name="DefaultTillageName">Default name of the tillage.</param>
@@ -1068,7 +1092,7 @@ namespace Models.Soils
             //-->
 
 
-            SurfaceOrganicMatter.TillageTypesList defaultTypes = new SurfaceOrganicMatter.TillageTypesList();
+            TillageTypesList defaultTypes = new TillageTypesList();
 
             TillageType data = defaultTypes.GetTillageData(DefaultTillageName);
 
@@ -1364,7 +1388,7 @@ namespace Models.Soils
         /// <summary>
         /// The surface
         /// </summary>
-        private Surface surface;
+        private Models.Soils.SoilWaterBackend.Surface surface;
 
         //Soil
         /// <summary>
